+++ clang/include/clang/ASTMatchers/ASTMatchers.h
@@ -310,6 +310,14 @@ AST_POLYMORPHIC_MATCHER_P(isExpansionInFileMatching,
 /// \endcode
 extern const internal::VariadicAllOfMatcher<Decl> decl;
 
+/// Matches decomposition-declarations.
+///
+/// Examples matches the declaration node.;
+/// \code
+///   auto [foo, bar] = std::make_pair{42, 42};
+/// \endcode
+extern const internal::VariadicAllOfMatcher<DecompositionDecl> decompositionDecl;
+
 /// Matches a declaration of a linkage specification.
 ///
 /// Given
@@ -4189,6 +4197,106 @@ AST_POLYMORPHIC_MATCHER_P2(forEachArgumentWithParam,
       }
     }
     ++ParamIndex;
+  }
+  *Builder = std::move(Result);
+  return Matched;
+}
+
+/// Matches all arguments and their respective Types. It is very similar to
+/// 'forEachArgumentWithParam' but it works on calls through function pointers
+/// as well. The difference is, that function pointers to not have 'ParmVarDecl'
+/// but only 'QualType' for each argument.
+///
+/// Given
+/// \code
+///   void f(int i);
+///   int y;
+///   f(y);
+///   void (*f_ptr)(int) = f;
+///   f_ptr(y);
+/// \endcode
+/// callExpr(
+///   forEachArgumentWithParamType(
+///     declRefExpr(to(varDecl(hasName("y")))),
+///     parmVarDecl(hasType(isInteger()))
+/// ))
+///   matches f(y);
+/// with declRefExpr(...)
+///   matching int y
+/// and parmVarDecl(...)
+///   matching int i
+AST_POLYMORPHIC_MATCHER_P2(forEachArgumentWithParamType,
+                           AST_POLYMORPHIC_SUPPORTED_TYPES(CallExpr,
+                                                           CXXConstructExpr),
+                           internal::Matcher<Expr>, ArgMatcher,
+                           internal::Matcher<QualType>, ParamMatcher) {
+  BoundNodesTreeBuilder Result;
+  // The first argument of an overloaded member operator is the implicit object
+  // argument of the method which should not be matched against a parameter, so
+  // we skip over it here.
+  BoundNodesTreeBuilder Matches;
+  unsigned ArgIndex = cxxOperatorCallExpr(callee(cxxMethodDecl()))
+                              .matches(Node, Finder, &Matches)
+                          ? 1
+                          : 0;
+
+  const FunctionProtoType *FProto = nullptr;
+
+  if (const auto *Call = dyn_cast<CallExpr>(&Node)) {
+    if (const Decl *Callee = Call->getCalleeDecl()) {
+      if (const auto *Value = dyn_cast<ValueDecl>(Callee)) {
+        QualType QT = Value->getType();
+
+        if (QT->isFunctionPointerType()) {
+          QualType FPT = QT->getPointeeType();
+          FProto = FPT->getAs<FunctionProtoType>();
+          assert(FProto &&
+                 "The call must have happened through a function pointer");
+        }
+
+        if (QT->isMemberFunctionPointerType()) {
+          const auto *MP = QT->getAs<MemberPointerType>();
+          assert(MP &&
+                 "Must be member-pointer if its a memberfunctionpointer");
+          FProto = MP->getPointeeType()->getAs<FunctionProtoType>();
+          assert(FProto &&
+                 "The call must have happened through a member function pointer");
+        }
+      }
+    }
+  }
+
+  int ParamIndex = 0;
+  bool Matched = false;
+
+  for (; ArgIndex < Node.getNumArgs(); ++ArgIndex, ++ParamIndex) {
+    BoundNodesTreeBuilder ArgMatches(*Builder);
+    if (ArgMatcher.matches(*(Node.getArg(ArgIndex)->IgnoreParenCasts()),
+                           Finder, &ArgMatches)) {
+      BoundNodesTreeBuilder ParamMatches(ArgMatches);
+
+      // This test is cheaper compared to the big matcher in the next if.
+      // Therefore, please keep this order.
+      if (FProto) {
+        QualType ParamType = FProto->getParamType(ParamIndex);
+        BoundNodesTreeBuilder ParamMatches(ArgMatches);
+
+        if (ParamMatcher.matches(ParamType, Finder, &ParamMatches)) {
+          Result.addMatch(ParamMatches);
+          Matched = true;
+          continue;
+        }
+      }
+      if (expr(anyOf(cxxConstructExpr(hasDeclaration(cxxConstructorDecl(
+                         hasParameter(ParamIndex, hasType(ParamMatcher))))),
+                     callExpr(callee(functionDecl(
+                         hasParameter(ParamIndex, hasType(ParamMatcher)))))))
+              .matches(Node, Finder, &ParamMatches)) {
+        Result.addMatch(ParamMatches);
+        Matched = true;
+        continue;
+      }
+    }
   }
   *Builder = std::move(Result);
   return Matched;